requirements.txt

absl-py==1.2.0
asgiref==3.5.2
astunparse==1.6.3
attrs==22.1.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
cheroot==8.6.0
CherryPy==18.8.0
decorator==5.1.1
distlib==0.3.5
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0


google-auth==2.10.0

google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
idna==3.3
importlib-metadata==4.12.0

jaraco.classes==3.2.2
jaraco.collections==3.5.2
jaraco.context==4.1.2
jaraco.functools==3.5.1
jaraco.text==3.8.1
joblib==1.1.0
jsonschema==4.9.1
keras==2.9.0
Keras-Preprocessing==1.1.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
more-itertools==8.14.0
numpy==1.23.1
oauthlib==3.2.0

opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
portend==3.1.0

pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.4.0
pymongo==4.2.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.6
python-slugify==6.1.2
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.0
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
tempora==5.0.2

tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.9.0
tensorflow==2.9.2
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==3.1.0

tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.2.2
wrapt==1.14.1
ws4py==0.5.1
zc.lockfile==2.0
zipp==3.8.1
